[Refactor - #53] DTO 패키지 분리로 인한 변경점
