+++ src/main/java/com/dateplan/dateplan/domain/member/service/AuthService.java
@@ -8,10 +8,10 @@ import static com.dateplan.dateplan.global.constant.Auth.SUBJECT_REFRESH_TOKEN;
 
 import com.dateplan.dateplan.domain.couple.service.CoupleReadService;
 import com.dateplan.dateplan.domain.member.dto.AuthToken;
-import com.dateplan.dateplan.domain.member.dto.LoginServiceResponse;
-import com.dateplan.dateplan.domain.member.dto.LoginServiceRequest;
-import com.dateplan.dateplan.domain.member.dto.PhoneAuthCodeServiceRequest;
-import com.dateplan.dateplan.domain.member.dto.PhoneServiceRequest;
+import com.dateplan.dateplan.domain.member.dto.login.LoginServiceResponse;
+import com.dateplan.dateplan.domain.member.dto.login.LoginServiceRequest;
+import com.dateplan.dateplan.domain.member.dto.signup.PhoneAuthCodeServiceRequest;
+import com.dateplan.dateplan.domain.member.dto.signup.PhoneServiceRequest;
 import com.dateplan.dateplan.domain.member.entity.Member;
 import com.dateplan.dateplan.domain.sms.service.SmsSendClient;
 import com.dateplan.dateplan.global.auth.JwtProvider;
